# perf: ECMAScript2023 &  ECMAScript2022 array methods performance improvement for proxy 