--- a/.changeset/great-trainers-beg.md
+++ b/.changeset/great-trainers-beg.md
@@ -1,0 +1,5 @@
+---
+"mobx": minor
+---
+
+Added es2023 array methods support to observablearray.