--- a/packages/mobx/__tests__/v5/base/array.js
+++ b/packages/mobx/__tests__/v5/base/array.js
@@ -857,6 +857,11 @@ describe("dehances", () => {
         expect(array.with(1, 5)).toEqual(dehanced.with(1, 5))
     })
 
+    test("at", () => {
+        expect(array.at(1)).toEqual(dehanced.at(1))
+        expect(array.at(-1)).toEqual(dehanced.at(-1))
+    })
+
     test("flat/flatMap", () => {
         const nestedArray = [{ value: 1 }, [{ value: 2 }, [{ value: 3 }]]]
         const dehancedNestedArray = nestedArray.map(dehancer)